Adjust the Gemfile in preparation for generating a static site
The https://git-scm.com/ site can easily be turned into a static site
because it does not really need any interactive features that require an
app on the server side.

So let's prepare for that by dropping all the unnecessary Ruby
dependencies.

We will add those back that are still needed, together with the commits
that introduce (or adapt) the code needing those dependencies. This way
we avoid keeping unneeded dependencies.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>