Respect the language in non-live search
When searching "live", thanks to the `lang` attribute of the `<html>`
element, Pagefind restricts the search to the current language. That is,
when looking at a French translation of a manual page, or at a page in
the French version of the ProGit2 book, typing a search term into the
search box will look only at the French pages of the site.

However, when pressing the Enter key, the browser navigates to
/search/results, whose `<html>` element is marked up with `lang=en`.
This is correct: the language of that page is English. Nevertheless, we
would want to see the same results (but this time with excerpts), i.e.
we want the search results page to respect the language of the
page from which the search originated.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>