Live search: only load 10 results at a time
There can be a lot of search results. Like, _really_ a lot. In that
case, it takes quite a bit of time to `await` all of the result data,
and that makes for a bad user experience.

Let's load them one at a time, and for good measure give more visual
indicators as to what is currently happening.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>