Allow `redirect_to` with fully-qualified URLs
So far, we have used the `redirect_to` directive in the front matter
only for redirecting to a relative URL.

However, in the next commit, we will want to add fall-backs for pages
that are not rendered on git-scm.com. This fall-back wants to redirect
to the git/git repository and hence does not need `relURL` treatment.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>